Rename RejectionSample() to be honest
RejectionSample() does not, in fact, rejection sample anything, despite
the numerous exhortations in the comments of various files that
rejection sampling must be used.

Because the functionality actually implemented by RejectionSample() was
literally just modulo, rename the function to reflect its implementation
more accurately.

This is okay, and has been okay, because the particular values of q used
with RejectionSample() happen to be ones sufficiently close to powers of
2 to not cause significant bias.