Add and verify factor proofs in resharing
Because of how the resharing is structured, it is necessary to add more
content to round 5. Round 4 is the earliest the new committee has access
to other member's H1, H2 and NTilde, which are necessary to produce
factor proofs for their own N and NTilde. That means that we need
another set of private messages in round 4, and the proofs in those
messages have to be verified in round 5.